Rename PubSubEventHandler classes to PubSubHandler
