+++ slack/adapter/src/commonMain/kotlin/com/gchristov/thecodinglove/slack/adapter/SlackAdapterModule.kt
@@ -10,12 +10,12 @@ import com.gchristov.thecodinglove.common.pubsub.PubSubDecoder
 import com.gchristov.thecodinglove.common.pubsub.PubSubPublisher
 import com.gchristov.thecodinglove.common.slack.SlackSender
 import com.gchristov.thecodinglove.slack.adapter.http.*
-import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackCancelSearchPubSubEventHandler
+import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackCancelSearchPubSubHandler
 import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackInteractivityPubSubHandler
 import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackSearchPubSubHandler
-import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackSelfDestructPubSubEventHandler
-import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackSendPubSubEventHandler
-import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackShufflePubSubEventHandler
+import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackSelfDestructPubSubHandler
+import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackSendPubSubHandler
+import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackShufflePubSubHandler
 import com.gchristov.thecodinglove.slack.adapter.search.RealSlackSearchRepository
 import com.gchristov.thecodinglove.slack.adapter.search.SlackSearchServiceApi
 import com.gchristov.thecodinglove.slack.domain.SlackMessageFactory
@@ -249,10 +249,10 @@ object SlackAdapterModule : DiModule() {
         jsonSerializer = jsonSerializer,
         log = log,
         eventHandlers = listOf(
-            SlackSendPubSubEventHandler(slackSendSearchUseCase, analytics),
-            SlackSelfDestructPubSubEventHandler(slackSendSearchUseCase, analytics),
-            SlackShufflePubSubEventHandler(slackShuffleSearchUseCase, analytics),
-            SlackCancelSearchPubSubEventHandler(slackCancelSearchUseCase, analytics),
+            SlackSendPubSubHandler(slackSendSearchUseCase, analytics),
+            SlackSelfDestructPubSubHandler(slackSendSearchUseCase, analytics),
+            SlackShufflePubSubHandler(slackShuffleSearchUseCase, analytics),
+            SlackCancelSearchPubSubHandler(slackCancelSearchUseCase, analytics),
         ),
         pubSubDecoder = pubSubDecoder,
     )
