+++ slack/adapter/src/commonMain/kotlin/com/gchristov/thecodinglove/slack/adapter/pubsub/SlackCancelSearchPubSubHandler.kt
@@ -8,12 +8,12 @@ import com.gchristov.thecodinglove.common.pubsub.PubSubDecoder
 import com.gchristov.thecodinglove.common.pubsub.PubSubHandler
 import com.gchristov.thecodinglove.slack.adapter.pubsub.model.SlackInteractivityReceivedEvent
 import com.gchristov.thecodinglove.slack.domain.model.SlackActionName
-import com.gchristov.thecodinglove.slack.domain.usecase.SlackSendSearchUseCase
+import com.gchristov.thecodinglove.slack.domain.usecase.SlackCancelSearchUseCase
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.serialization.DeserializationStrategy
 
-internal class SlackSelfDestructPubSubEventHandler(
-    private val slackSendSearchUseCase: SlackSendSearchUseCase,
+internal class SlackCancelSearchPubSubHandler(
+    private val slackCancelSearchUseCase: SlackCancelSearchUseCase,
     private val analytics: Analytics,
 ) : PubSubHandler<SlackInteractivityReceivedEvent.InteractivityPayload.InteractiveMessage> {
     override val dispatcher: CoroutineDispatcher get() = error("not used")
@@ -24,21 +24,17 @@ internal class SlackSelfDestructPubSubEventHandler(
     override fun httpConfig() = error("not used")
 
     override suspend fun handle(event: SlackInteractivityReceivedEvent.InteractivityPayload.InteractiveMessage): Either<Throwable, Unit> {
-        val action = event.actions.firstOrNull { it.name == SlackActionName.SELF_DESTRUCT_5_MIN.apiValue }
+        val action = event.actions.firstOrNull { it.name == SlackActionName.CANCEL.apiValue }
             ?: return Either.Right(Unit)
         analytics.sendEvent(
             clientId = event.user.id,
-            name = "slack_interactivity_self_destruct",
+            name = "slack_interactivity_cancel",
             params = mapOf("user_id" to event.user.id, "team_id" to event.team.id),
         )
-        return slackSendSearchUseCase(
-            SlackSendSearchUseCase.Dto(
-                userId = event.user.id,
-                teamId = event.team.id,
-                channelId = event.channel.id,
+        return slackCancelSearchUseCase(
+            SlackCancelSearchUseCase.Dto(
                 responseUrl = event.responseUrl,
                 searchSessionId = action.value,
-                selfDestructMinutes = 5,
             )
         )
     }
