+++ slack/adapter/src/commonMain/kotlin/com/gchristov/thecodinglove/slack/adapter/pubsub/SlackSendPubSubHandler.kt
@@ -8,12 +8,12 @@ import com.gchristov.thecodinglove.common.pubsub.PubSubDecoder
 import com.gchristov.thecodinglove.common.pubsub.PubSubHandler
 import com.gchristov.thecodinglove.slack.adapter.pubsub.model.SlackInteractivityReceivedEvent
 import com.gchristov.thecodinglove.slack.domain.model.SlackActionName
-import com.gchristov.thecodinglove.slack.domain.usecase.SlackShuffleSearchUseCase
+import com.gchristov.thecodinglove.slack.domain.usecase.SlackSendSearchUseCase
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.serialization.DeserializationStrategy
 
-internal class SlackShufflePubSubEventHandler(
-    private val slackShuffleSearchUseCase: SlackShuffleSearchUseCase,
+internal class SlackSendPubSubHandler(
+    private val slackSendSearchUseCase: SlackSendSearchUseCase,
     private val analytics: Analytics,
 ) : PubSubHandler<SlackInteractivityReceivedEvent.InteractivityPayload.InteractiveMessage> {
     override val dispatcher: CoroutineDispatcher get() = error("not used")
@@ -24,17 +24,21 @@ internal class SlackShufflePubSubEventHandler(
     override fun httpConfig() = error("not used")
 
     override suspend fun handle(event: SlackInteractivityReceivedEvent.InteractivityPayload.InteractiveMessage): Either<Throwable, Unit> {
-        val action = event.actions.firstOrNull { it.name == SlackActionName.SHUFFLE.apiValue }
+        val action = event.actions.firstOrNull { it.name == SlackActionName.SEND.apiValue }
             ?: return Either.Right(Unit)
         analytics.sendEvent(
             clientId = event.user.id,
-            name = "slack_interactivity_shuffle",
+            name = "slack_interactivity_send",
             params = mapOf("user_id" to event.user.id, "team_id" to event.team.id),
         )
-        return slackShuffleSearchUseCase(
-            SlackShuffleSearchUseCase.Dto(
+        return slackSendSearchUseCase(
+            SlackSendSearchUseCase.Dto(
+                userId = event.user.id,
+                teamId = event.team.id,
+                channelId = event.channel.id,
                 responseUrl = event.responseUrl,
                 searchSessionId = action.value,
+                selfDestructMinutes = null,
             )
         )
     }
